requirements.txt
GPUtil==1.4.0
numpy==2.2.1
PyYAML==6.0.2
scikit_learn==1.6.0
torch==2.5.1
transformers==4.42.4
